Fix single-arg `Coordinate()` constructor
`Coordinate()` has a single-arg (not counting the `self` arg)
constructor overload that is unusable because the overloaded constructor
requires two args.

    from stgpytools import Coordinate

    Coordinate((0, 0))
    # TypeError: Coordinate.__init__() missing 1 required positional argument: 'y'

This is caught by mypy but the error was being `type: ignore`'d. This
patch fixes the issue by making the `y` arg optional in the overloaded
constructor's signature (but it's still required overall).

This was originally discovered via vs-masktool's `BoundingBox`:

    from vsmasktools import BoundingBox

    BoundingBox((0, 0), (100, 100))
    #   File "vs-masktools/vsmasktools/abstract.py", line 36, in __init__
    #     self.pos, self.size, self.invert = Position(pos), Size(size), invert
    #                                        ^^^^^^^^^^^^^
    # TypeError: Coordinate.__init__() missing 1 required positional argument: 'y'