CustomEnum: Release infinite recursion (#11)
Currently, an issue occurs if you pass i.e. Matrix(-1). It will continually call `cls(value)`, which in turn calls `cls.from_param`, which eventually circles back to `_missing_`.

This at least puts a stopgap to this issue and properly raises the error, although I'm sure there's a better solution that fundamentally deals with the problem.